+++ src/theme.rs
@@ -1,5 +1,5 @@
+use iced::Theme;
 use iced::theme::Palette;
-use iced::{Color, Theme};
 
 use crate::styles;
 
@@ -7,11 +7,7 @@ pub fn default() -> Theme {
     Theme::custom(
         "provision",
         Palette {
-            background: Color::from_rgb(
-                0x09 as f32 / 255.0,
-                0x09 as f32 / 255.0,
-                0x0b as f32 / 255.0,
-            ), // zinc-950
+            background: styles::BG,
             text: styles::TEXT,
             primary: styles::STATUS_BLUE,
             success: styles::STATUS_GREEN,